add oscillation precision
The ocsillation feature only turns on if the missile is within a forward facing angle towards the target -- basically meaning when it's behind it's target, and not turning outisde of a cone that could be visually cast in front of it.

The oscillator is made to send the first oscillation in the same direction that the missile is already traveling to attempt to prevent it getting stuck on the edge of the precision cone. If it enters from the left and the oscillator pulses a slight turn to the left, it will knock it back out of the precision cone, and another frame goes by, then back in for one frame and back out and so on.